reflex-binance/requirements.txt
reflex==0.2.0
python-dotenv
python-binance
